// src/lib/auth/api/api.ts
export interface V1AdminRolesGet200Response {
    /**
     * 
     * @type {string}
     * @memberof V1AdminRolesGet200Response
     */
    'code': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminRolesGet200Response
     */
    'message': string | null;
    /**
     * 
     * @type {V1AdminRolesGet200ResponseData}
     * @memberof V1AdminRolesGet200Response
     */
    'data': V1AdminRolesGet200ResponseData;
}
/**
 * 
 * @export
 * @interface V1AdminRolesGet200ResponseData
 */
export interface V1AdminRolesGet200ResponseData {
    /**
     * 
     * @type {Array<V1AdminRolesGet200ResponseDataDataInner>}
     * @memberof V1AdminRolesGet200ResponseData
     */
    'data': Array<V1AdminRolesGet200ResponseDataDataInner>;
    /**
     * 
     * @type {number}
     * @memberof V1AdminRolesGet200ResponseData
     */
    'count': number;
}
/**
 * 
 * @export
 * @interface V1AdminRolesGet200ResponseDataDataInner
 */
export interface V1AdminRolesGet200ResponseDataDataInner {
    /**
     * 
     * @type {string}
     * @memberof V1AdminRolesGet200ResponseDataDataInner
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminRolesGet200ResponseDataDataInner
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminRolesGet200ResponseDataDataInner
     */
    'updatedAt': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminRolesGet200ResponseDataDataInner
     */
    'deletedAt'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminRolesGet200ResponseDataDataInner
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminRolesGet200ResponseDataDataInner
     */
    'description'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof V1AdminRolesGet200ResponseDataDataInner
     */
    'numUsers'?: number;
}
// ...
export interface V1AdminUsersGet200Response {
    /**
     * 
     * @type {string}
     * @memberof V1AdminUsersGet200Response
     */
    'code': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminUsersGet200Response
     */
    'message': string | null;
    /**
     * 
     * @type {V1AdminUsersGet200ResponseData}
     * @memberof V1AdminUsersGet200Response
     */
    'data': V1AdminUsersGet200ResponseData;
}
/**
 * 
 * @export
 * @interface V1AdminUsersGet200ResponseData
 */
export interface V1AdminUsersGet200ResponseData {
    /**
     * 
     * @type {Array<V1AdminUsersIdGet200ResponseDataData>}
     * @memberof V1AdminUsersGet200ResponseData
     */
    'data': Array<V1AdminUsersIdGet200ResponseDataData>;
    /**
     * 
     * @type {number}
     * @memberof V1AdminUsersGet200ResponseData
     */
    'count': number;
}